+++ unstd/src/picking/mod.rs
@@ -6,6 +6,7 @@
 
 mod gamepad_pointer;
 mod sprite_picking_backend;
+mod truck_nav;
 
 pub use gamepad_pointer::GamepadPointerPlugin;
 pub use sprite_picking_backend::{